+++ packages/redis/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./store";
-export * from "./types";
+export type * from "./types";