+++ packages/redis/src/store.ts
@@ -3,6 +3,7 @@ import type {
   ConfigType as RateLimitConfiguration,
   Store,
 } from "hono-rate-limiter";
+import type { Env, Input } from "hono/types";
 import scripts from "./scripts";
 import type { Options, RedisClient, RedisReply } from "./types";
 
@@ -38,7 +39,12 @@ const parseScriptResponse = (results: RedisReply): ClientRateLimitInfo => {
   return { totalHits, resetTime };
 };
 
-export class RedisStore implements Store {
+export class RedisStore<
+  E extends Env = Env,
+  P extends string = string,
+  I extends Input = Input,
+> implements Store<E, P, I>
+{
   /**
    * The text to prepend to the key in Redis.
    */
@@ -119,7 +125,6 @@ export class RedisStore implements Store {
     const evalCommand = async () =>
       this.client.evalsha<string[], RedisReply>(
         await this.incrementScriptSha,
-        // "1",
         [this.prefixKey(key)],
         [this.resetExpiryOnChange ? "1" : "0", this.windowMs.toString()],
       );
@@ -150,7 +155,7 @@ export class RedisStore implements Store {
    *
    * @param options {RateLimitConfiguration} - The options used to setup the middleware.
    */
-  init(options: RateLimitConfiguration) {
+  init(options: RateLimitConfiguration<E, P, I>) {
     this.windowMs = options.windowMs;
   }